chore(deps): `go mod`: update

+++ data/roadmap/examples/roadmap/roadmap.go
@@ -12,14 +12,14 @@ import (
 )
 
 func main() {
-	can, err := roadmap.GetCanvasQuarter(int32(20174), int32(20182))
+	can, err := roadmap.GetCanvasQuarter(20174, 20182)
 	if err != nil {
 		slog.Error(err.Error())
 		os.Exit(1)
 	}
 
 	item := roadmap.Item{Name: "Feature 1"}
-	err = item.SetMinMaxQuarter(int32(20174), int32(20174))
+	err = item.SetMinMaxQuarter(20174, 20174)
 	if err != nil {
 		slog.Error(err.Error())
 		os.Exit(2)
@@ -27,7 +27,7 @@ func main() {
 	can.AddItem(item)
 
 	item = roadmap.Item{Name: "Feature 2"}
-	err = item.SetMinMaxQuarter(int32(20174), int32(20181))
+	err = item.SetMinMaxQuarter(20174, 20181)
 	if err != nil {
 		slog.Error(err.Error())
 		os.Exit(3)
@@ -41,7 +41,7 @@ func main() {
 	}
 
 	item2 := roadmap.Item{Name: "Feature 3"}
-	err = item2.SetMinMaxQuarter(int32(20181), int32(20182))
+	err = item2.SetMinMaxQuarter(20181, 20182)
 	if err != nil {
 		slog.Error(err.Error())
 		os.Exit(4)
